+++ Makefile.am
@@ -1074,8 +1074,8 @@ libarchive_test_EXTRA_DIST=\
 # Common code for libarchive frontends (cpio, tar)
 #
 libarchive_fe_la_SOURCES= \
-	libarchive_fe/err.c \
-	libarchive_fe/err.h \
+	libarchive_fe/lafe_err.c \
+	libarchive_fe/lafe_err.h \
 	libarchive_fe/lafe_platform.h \
 	libarchive_fe/line_reader.c \
 	libarchive_fe/line_reader.h \
