+++ cat/CMakeLists.txt
@@ -13,8 +13,8 @@ IF(ENABLE_CAT)
     bsdcat.h
     bsdcat_platform.h
     cmdline.c
-    ../libarchive_fe/err.c
-    ../libarchive_fe/err.h
+    ../libarchive_fe/lafe_err.c
+    ../libarchive_fe/lafe_err.h
     ../libarchive_fe/lafe_platform.h
   )
   INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/../libarchive_fe)
