+++ contrib/android/Android.mk
@@ -151,7 +151,7 @@ else
 libarchive_host_src_files :=
 endif
 
-libarchive_fe_src_files :=  libarchive_fe/err.c \
+libarchive_fe_src_files :=  libarchive_fe/lafe_err.c \
 							libarchive_fe/line_reader.c \
 							libarchive_fe/passphrase.c
