+++ cpio/CMakeLists.txt
@@ -13,8 +13,8 @@ IF(ENABLE_CPIO)
     cpio.c
     cpio.h
     cpio_platform.h
-    ../libarchive_fe/err.c
-    ../libarchive_fe/err.h
+    ../libarchive_fe/lafe_err.c
+    ../libarchive_fe/lafe_err.h
     ../libarchive_fe/lafe_platform.h
     ../libarchive_fe/line_reader.c
     ../libarchive_fe/line_reader.h
