+++ tar/CMakeLists.txt
@@ -18,8 +18,8 @@ IF(ENABLE_TAR)
     subst.c
     util.c
     write.c
-    ../libarchive_fe/err.c
-    ../libarchive_fe/err.h
+    ../libarchive_fe/lafe_err.c
+    ../libarchive_fe/lafe_err.h
     ../libarchive_fe/lafe_platform.h
     ../libarchive_fe/line_reader.c
     ../libarchive_fe/line_reader.h
