+++ unzip/CMakeLists.txt
@@ -16,8 +16,8 @@ IF(ENABLE_UNZIP)
     la_getline.c
     la_getline.h
     la_queue.h
-    ../libarchive_fe/err.c
-    ../libarchive_fe/err.h
+    ../libarchive_fe/lafe_err.c
+    ../libarchive_fe/lafe_err.h
     ../libarchive_fe/lafe_platform.h
     ../libarchive_fe/passphrase.c
     ../libarchive_fe/passphrase.h
